% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult_ringed_seal}
\alias{adult_ringed_seal}
\title{Track data from the adult ringed seal with tag id 106373.}
\format{A data frame with 2698 rows and 5 variables:
\describe{
  \item{id}{Tag id}
  \item{date}{Date and time of the observation}
  \item{lc}{Argos location class of the observation}
  \item{lat}{Observed latitude (degrees north) from the Argos system}
  \item{lon}{Observed longitude (degrees east) from the Argos system}

}}
\source{
Fisheries and Oceans Canada
}
\usage{
adult_ringed_seal
}
\description{
A dataset containing tracking data from the Argos system for an adult ringed seal (Pusa hispida) released 10-30-2011 near Sanikiluaq, Nunavut, Canada.
}
\keyword{datasets}

