% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Animal.R
\name{Animal}
\alias{Animal}
\title{Construct an Animal object}
\usage{
Animal(observation, movement, measurement, name = "", ...)
}
\arguments{
\item{observation}{an Observation object}

\item{movement}{a Movement object}

\item{measurement}{a Measurement object}

\item{name}{Name of the animal}

\item{...}{Other arguments not currently used}
}
\value{
An Animal class object
}
\description{
Construct an Animal object
}
\examples{
d <- subadult_ringed_seal
obs <- Observation(lon = d$lon,
                   lat = d$lat,
                   dates = as.POSIXct(d$date,tz="GMT"),
                   locationclass = d$lc
                   )
meas <- Measurement(model="n")
mov <- RW(unique(obs$dates))
anim <- Animal(obs,mov,meas,"Subadult")
plot(anim)
}
\seealso{
\code{\link{Animal-class}} \code{\link{Movement-class}} \code{\link{Observation-class}} \code{\link{Measurement-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
