% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/argostrack.R
\docType{package}
\name{argosTrack-package}
\alias{argosTrack-package}
\title{Estimate movement models to Argos data.}
\description{
argosTrack allows you to fit different movement models to data.
}
\details{
The main function in \pkg{argosTrack} is
\code{\link{argosTrack}}. The package also provides plot, print summary and simulation functions.
}

