% Generated by roxygen2 (4.0.2): do not edit by hand
\name{argosTrack}
\alias{argosTrack}
\title{Estimate track from Argos data}
\usage{
argosTrack(lon, lat, dates, locationclass, include = rep(TRUE, length(dates)),
  equalbetas = TRUE, fixgammas = TRUE, df = 10, errordistribution = "t",
  verbose = TRUE)
}
\arguments{
\item{lon}{}

\item{lat}{}

\item{dates}{}

\item{locationclass}{}

\item{include}{}

\item{equalbetas}{}

\item{fixgammas}{}

\item{df}{}

\item{errordistribution}{}

\item{verbose}{Write maximum gradient components to the terminal?}
}
\description{
Estimate animal movement track from Argos data using TMB (tmb-project.org).
}

