% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Movement_DCRW.R
\docType{class}
\name{DCRW-class}
\alias{DCRW}
\alias{DCRW-class}
\title{A Reference Class for fitting a discrete time first Difference Correlated Random Walk model.}
\description{
The reference class implements a discrete time first Difference Correlated Random Walk (Jonsen et al. 2005). The locations are modelled by
\deqn{X_t = X_{t-\Delta} + \gamma T(\phi) (X_t - X_{t-\Delta}) + \epsilon_t}
Here, \eqn{\epsilon_t} is zero mean Gaussian noise with covariance \eqn{ \pmatrix{ \sigma_1^2 & \rho\sigma_1\sigma_2 \cr \rho\sigma_1\sigma_2 & \sigma_2^2 }}. \eqn{T(\phi)} is the rotation matrix \eqn{ \pmatrix{ \cos(\phi) & -\sin(\phi) \cr \sin(\phi) & \cos(\phi) }}. \eqn{\gamma} is a scalar.
}
\section{Methods}{

\describe{
\item{\code{initialize(dates, pars = numeric(3), varPars = numeric(2),
  nauticalStates = FALSE, timeunit = "hours")}}{Method to initialize the class. 'dates' is a vector of distinct and equidistant POSIXct dates; 'pars' is vector of the movement parameters: \eqn{logit_{(0,1)}(\gamma)}, \eqn{\phi}, \eqn{logit_{(-1,1)}(\rho)}; 'varPars' is a vector of movement variance parameters: \eqn{log(\sigma_{lat})}, \eqn{log(\sigma_{lat})}; 'nauticalStates' is a logical value indicating whether the states should be modelled in nautical miles, and 'timeunit' is the time unit to use for calculating time steps.}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}
\examples{
d <- subadult_ringed_seal
dates <- unique(as.POSIXct(d$date,tz="GMT"))
dseq <- seq(min(dates),max(dates), "day")
mov <- DCRW(dseq)
}
\references{
Jonsen, I., J. Mills Flemming, and R. Myers. (2005) Robust state-space modeling of animal movement data. Ecology 86, 2874-2880.
}
\seealso{
\link{Movement}

Other X.Movement.models.: \code{\link{CSB}},
  \code{\link{CSB-class}}; \code{\link{CTCRW}},
  \code{\link{CTCRW-class}}; \code{\link{DIRAC}},
  \code{\link{DIRAC-class}}; \code{\link{DSBHN}},
  \code{\link{DSBHN-class}}; \code{\link{DSBW}},
  \code{\link{DSBW-class}}; \code{\link{IDCRW}},
  \code{\link{IDCRW-class}}; \code{\link{MPCTCRW}},
  \code{\link{MPCTCRW-class}}; \code{\link{Movement}},
  \code{\link{Movement-class}}; \code{\link{OUL}},
  \code{\link{OUL-class}}; \code{\link{OUV}},
  \code{\link{OUV-class}}; \code{\link{RW}},
  \code{\link{RW-class}}
}

