% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_DCRW.R
\name{DCRW}
\alias{DCRW}
\title{Create a DCRW movement model object}
\usage{
DCRW(
  dates,
  pars = numeric(3),
  varPars = numeric(2),
  nauticalStates = FALSE,
  timeunit = "hours"
)
}
\arguments{
\item{dates}{Vector of distinct, equidistant, and increasing POSIXct dates}

\item{pars}{Vector of movement parameters: \\eqn{logit_{(0,1)}(\\gamma)}, \\eqn{\\phi}, \\eqn{logit_{(-1,1)}(\\rho)}}

\item{varPars}{Vector of movement variance parameters: \\eqn{log(\\sigma_{lat})}, \\eqn{log(\\sigma_{lat})}}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
A DCRW object
}
\description{
Create a DCRW movement model object
}
\examples{
d <- subadult_ringed_seal
dates <- unique(as.POSIXct(d$date,tz="GMT"))
dseq <- seq(min(dates),max(dates), "day")
mov <- DCRW(dseq)
}
\seealso{
\code{\link{DCRW-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
