% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Movement_DIRAC.R
\docType{class}
\name{DIRAC-class}
\alias{DIRAC}
\alias{DIRAC-class}
\title{A Reference Class for fitting a Dirac movement model.}
\description{
The reference class implements a Dirac movement model with no movement
\deqn{X_t \sim \delta(X_0))}
}
\section{Methods}{

\describe{
\item{\code{getTMBmap(...)}}{Function to return a map list for TMB::MakeADFun.}

\item{\code{initialize(dates, pars = numeric(2), varPars = numeric(0),
  nauticalStates = FALSE, timeunit = "hours")}}{Method to initialize the class. 'dates' is a vector of distinct and increasing POSIXct dates; 'pars' is a zero length vector of the movement parameters; 'varPars' is a zero length vector of movement variance parameters; 'nauticalStates' is a logical value indicating whether the states should be modelled in nautical miles, and 'timeunit' is the time unit to use for calculating time steps.}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. The function only returns the locations. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}
\examples{
d <- subadult_ringed_seal
mov <- argosTrack:::DIRAC(unique(as.POSIXct(d$date,tz="GMT")))
}
\seealso{
\link{Movement}

Other X.Movement.models.: \code{\link{CSB}},
  \code{\link{CSB-class}}; \code{\link{CTCRW}},
  \code{\link{CTCRW-class}}; \code{\link{DCRW}},
  \code{\link{DCRW-class}}; \code{\link{DSBHN}},
  \code{\link{DSBHN-class}}; \code{\link{DSBW}},
  \code{\link{DSBW-class}}; \code{\link{IDCRW}},
  \code{\link{IDCRW-class}}; \code{\link{MPCTCRW}},
  \code{\link{MPCTCRW-class}}; \code{\link{Movement}},
  \code{\link{Movement-class}}; \code{\link{OUL}},
  \code{\link{OUL-class}}; \code{\link{OUV}},
  \code{\link{OUV-class}}; \code{\link{RW}},
  \code{\link{RW-class}}
}

