% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_DIRAC.R
\name{DIRAC}
\alias{DIRAC}
\alias{DIRAC}
\title{A Reference Class for fitting a Dirac movement model.}
\usage{
DIRAC(dates, pars = numeric(0), varPars = numeric(0),
  nauticalStates = FALSE, timeunit = "hours")

DIRAC(dates, pars = numeric(0), varPars = numeric(0),
  nauticalStates = FALSE, timeunit = "hours")
}
\arguments{
\item{dates}{Vector of distinct and increasing POSIXct dates}

\item{pars}{Zero length vector of movement parameters}

\item{varPars}{Zero length vector of movement variance parameters}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
A DIRAC object
}
\description{
The reference class implements a Dirac movement model with no movement
\deqn{X_t \sim \delta(X_0))}

Create a DIRAC movement model object
}
\examples{
d <- subadult_ringed_seal
mov <- DIRAC(unique(as.POSIXct(d$date,tz="GMT")))

d <- subadult_ringed_seal
mov <- DIRAC(unique(as.POSIXct(d$date,tz="GMT")))
}
\seealso{
\link{Movement-class} \link{DIRAC}

\code{\link{DIRAC-class}}

Other "Movement models": \code{\link{CSB-class}},
  \code{\link{CTCRW-class}}, \code{\link{DCRW-class}},
  \code{\link{DSBHN-class}}, \code{\link{DSBW}},
  \code{\link{GDCRW-class}}, \code{\link{MPCTCRW-class}},
  \code{\link{Movement-class}}, \code{\link{OUL-class}},
  \code{\link{OUV-class}}, \code{\link{RW-class}}
}
\author{
Christoffer Moesgaard Albertsen

Christoffer Moesgaard Albertsen
}
