% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Movement_DSBHN.R
\docType{class}
\name{DSBHN-class}
\alias{DSBHN}
\alias{DSBHN-class}
\title{A Reference Class for fitting a Discrete time Step-length/Bearings model with Half Normal distributed steps.}
\description{
The reference class implements a Discrete time Step-length/Bearings model with Half Normal distributed steps. \cr
Let \eqn{X_t} be the (bivariate) location (in nautical miles) at time \eqn{t}. The step lengths are
\deqn{S_t = \sqrt{\sum\left(X_t-X_{t-1}\right)^2}}
and the bearings are
\deqn{B_t = atan2(X_{y,t}-X_{y,t-1},X_{x,t}-X_{y,t-1})}
where \eqn{x} denotes the east/west coordinate and \eqn{y} denotes the north/south coordinate.
The step lengths are modelled by a half normal distribution
\deqn{S_t \sim HN(\sigma^2)}
and the bearings are modelled by a wrapped Cauchy distribution with location parameter \eqn{B_{t-1}} and concentration parameter \eqn{\rho}.
}
\section{Methods}{

\describe{
\item{\code{getTMBparameters()}}{Function to return a parameter list for TMB::MakeADFun}

\item{\code{initialize(dates, pars = numeric(1), varPars = numeric(1),
  nauticalStates = FALSE, timeunit = "hours")}}{Method to initialize the class. 'dates' is a vector of distinct and equidistant POSIXct dates; 'pars' is vector of the movement parameters: \eqn{log(\rho)}; 'varPars' is a vector of movement variance parameters: \eqn{log(\sigma)}; 'nauticalStates' is a logical value indicating whether the states should be modelled in nautical miles, and 'timeunit' is the time unit to use for calculating time steps.}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}
\note{
Step-length/Bearings models can be difficult for the Laplace approximation. Estimation problems can often be solved by decreasing the number of states relative to the number of observations.
}
\examples{
d <- subadult_ringed_seal
dates <- unique(as.POSIXct(d$date))
dseq <- seq(min(dates),max(dates), "day")
mov <- DSBHN(dseq)
}
\seealso{
\code{\link{Movement}}, \code{link{DSBW}}.

Other X.Movement.models.: \code{\link{CTCRW}},
  \code{\link{CTCRW-class}}; \code{\link{DCRW}},
  \code{\link{DCRW-class}}; \code{\link{DSBW}},
  \code{\link{DSBW-class}}; \code{\link{IDCRW}},
  \code{\link{IDCRW-class}}; \code{\link{MPCTCRW}},
  \code{\link{MPCTCRW-class}}; \code{\link{Movement}},
  \code{\link{Movement-class}}; \code{\link{OUL}},
  \code{\link{OUL-class}}; \code{\link{RW}},
  \code{\link{RW-class}}
}

