% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_DSBHN.R
\name{DSBHN}
\alias{DSBHN}
\title{Create a DSBHN movement model object}
\usage{
DSBHN(dates, pars = numeric(1), varPars = numeric(1),
  nauticalStates = FALSE, timeunit = "hours")
}
\arguments{
\item{dates}{Vector of distinct and increasing POSIXct dates}

\item{pars}{Vector of movement parameters: \\eqn{log(\\rho)}}

\item{varPars}{Vector of movement variance parameters: \\eqn{log(\\sigma)}}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
A DSBHN object
}
\description{
Create a DSBHN movement model object
}
\examples{
d <- subadult_ringed_seal
dates <- unique(as.POSIXct(d$date,tz="GMT"))
dseq <- seq(min(dates),max(dates), "day")
mov <- DSBHN(dseq)
}
\seealso{
\code{\link{DSBHN-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
