% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_DSBW.R
\docType{class}
\name{DSBW-class}
\alias{DSBW-class}
\title{A Reference Class for fitting a Discrete time Step-length/Bearings model with Weibull distributed steps.}
\description{
The reference class implements a Discrete time Step-length/Bearings model with Weibull distributed steps (McClintock et al. 2015, with small modifications). \cr
Let \eqn{X_t} be the (bivariate) location (in nautical miles) at time \eqn{t}. The step lengths are
\deqn{S_t = \sqrt{\sum\left(X_t-X_{t-1}\right)^2}}
and the bearings are
\deqn{B_t = atan2(X_{y,t}-X_{y,t-1},X_{x,t}-X_{y,t-1})} 
where \eqn{x} denotes the east/west coordinate and \eqn{y} denotes the north/south coordinate.
The step lengths are modelled by a half normal distribution
\deqn{S_t \sim Weibull(scale,shape)}
and the bearings are modelled by a wrapped Cauchy distribution with location parameter \eqn{B_{t-1}} and concentration parameter \eqn{rho}.
}
\section{Methods}{

\describe{
\item{\code{getTMBparameters()}}{Function to return a parameter list for TMB::MakeADFun}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}

\note{
Step-length/Bearings models can be difficult for the Laplace approximation. Estimation problems can often be solved by using \code{\link{DSBHN}} instead or by decreasing the number of states relative to the number of observations.
}
\references{
McClintock, B. T., London, J. M., Cameron, M. F. and Boveng, P. L. (2015) Modelling animal movement using the Argos satellite telemetry location error ellipse. Methods Ecol Evol, 6: 266-277. doi:10.1111/2041-210X.12311
}
\seealso{
\code{\link{Movement-class}}, \code{\link{DSBW}}, \code{\link{DSBHN}}.

Other "Movement models": \code{\link{CSB-class}},
  \code{\link{CTCRW-class}}, \code{\link{DCRW-class}},
  \code{\link{DIRAC-class}}, \code{\link{DSBHN-class}},
  \code{\link{GDCRW-class}}, \code{\link{MPCTCRW-class}},
  \code{\link{Movement-class}}, \code{\link{OUL-class}},
  \code{\link{OUV-class}}, \code{\link{RW-class}}
}
