% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\name{expandMu}
\alias{expandMu}
\title{Expand estimated states to all time points}
\usage{
expandMu(eststat, dt)
}
\arguments{
\item{eststat}{2xn Matrix of estimated states}

\item{dt}{vector of length m>=n of time difference to previous observation}
}
\value{
matrix of expanded estimates
}
\description{
Expand estimated states to all observations
}
\details{
Internal function to expand estimated states to all observations when there are multiple observations at time points
}
\author{
Christoffer Moesgaard Albertsen
}
\keyword{internal}

