% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTrack.R
\name{fitTrack,Animal-method}
\alias{fitTrack,Animal-method}
\title{Fits a state-space movement model to animal data}
\usage{
\S4method{fitTrack}{Animal}(
  object,
  method = c("laplace", "mcmc"),
  silent = FALSE,
  fixcorrection = FALSE,
  nlminb.control = list(),
  inner.control = list(maxit = 1000),
  ...
)
}
\arguments{
\item{object}{\link{Animal} reference class object}

\item{method}{Estimation method. Either laplace for maximum likelihood estimation with the Laplace approximation or mcmc for Markov Chain Monte Carlo (not yet implemented).}

\item{silent}{Disable all tracing information?}

\item{fixcorrection}{Should the ratio between measurement errors be fixed?}

\item{nlminb.control}{List of control arguments passed to 'nlminb'}

\item{inner.control}{List of control arguments passed to 'TMB::newton'}

\item{...}{Additional arguments passed to object$getTMBmap}
}
\value{
The result returned from 'nlminb' is returned invisibly.
}
\description{
Fits the state-space movement determined by object using maximum likelihood estimation with the Laplace approximation.
}
\note{
The function changes 'object'.
}
\seealso{
\link[TMB]{MakeADFun} \link[TMB]{newton} \link[stats]{nlminb}

\code{\link{fitTrack}}
}
\author{
Christoffer Moesgaard Albertsen
}
