% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTrack.R
\name{fitTrack}
\alias{fitTrack}
\title{Generic function to fit movement models.}
\usage{
fitTrack(object, ...)
}
\arguments{
\item{object}{The object to be fitted}

\item{...}{Additional arguments passed to the appropriate method}
}
\value{
The optimization result
}
\description{
Generic function to fit movement models.
}
\seealso{
\code{\link{fitTrack,Animal-method}}
}
