% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_GDCRW.R
\name{GDCRW}
\alias{GDCRW}
\title{Create a GDCRW movement model object}
\usage{
GDCRW(dates, pars = numeric(6), varPars = numeric(2),
  nauticalStates = FALSE, timeunit = "hours")
}
\arguments{
\item{dates}{Vector of distinct and increasing POSIXct dates}

\item{pars}{Vector of movement parameters: \\eqn{logit_{(0,1)}(\\gamma_{lat})}, \\eqn{logit_{(0,1)}(gamma_{lon})}, \\eqn{\\phi}, \\eqn{logit_{(-1,1)}(\\rho)}, \\eqn{\\mu_{lat}}, \\eqn{\\mu_{lon}}}

\item{varPars}{Vector of movement variance parameters: \\eqn{log(\\sigma_{lat})}, \\eqn{log(\\sigma_{lon})}}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
A GDCRW object
}
\description{
Create a GDCRW movement model object
}
\examples{
d <- subadult_ringed_seal
mov <- GDCRW(unique(as.POSIXct(d$date,tz="GMT")))
}
\seealso{
\code{\link{GDCRW-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
