% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTrack.R
\docType{methods}
\name{getTrack,Animal-method}
\alias{getTrack,Animal-method}
\title{Extracts track from animal data}
\usage{
\S4method{getTrack}{Animal}(object, ...)
}
\arguments{
\item{object}{\link{Animal} reference class object}

\item{...}{Additional arguments passed to the appropriate method}
}
\value{
A data frame of positions
}
\description{
Extracts the estimated and observed track from an Animal object
}
\author{
Christoffer Moesgaard Albertsen
}

