% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrack.R
\docType{methods}
\name{getTrack,Movement-method}
\alias{getTrack,Movement-method}
\title{Extracts track from animal data}
\usage{
\S4method{getTrack}{Movement}(object, ...)
}
\arguments{
\item{object}{\link{Movement} reference class object}

\item{...}{Additional arguments passed to the appropriate method}
}
\value{
A data frame of positions
}
\description{
Extracts the estimated track from a Movement object
}
\author{
Christoffer Moesgaard Albertsen
}
