% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrack.R
\name{getTrack,Observation-method}
\alias{getTrack,Observation-method}
\title{Extracts track from animal data}
\usage{
\S4method{getTrack}{Observation}(object, ...)
}
\arguments{
\item{object}{\link{Observation} reference class object}

\item{...}{Additional arguments passed to the appropriate method}
}
\value{
A data frame of positions
}
\description{
Extracts the observed track from an Observation object
}
\author{
Christoffer Moesgaard Albertsen
}
