% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrack.R
\name{getTrack}
\alias{getTrack}
\title{Generic function to extract movement track.}
\usage{
getTrack(object, ...)
}
\arguments{
\item{object}{The object from which to extract the track}

\item{...}{Additional arguments passed to the appropriate method}
}
\value{
A data frame of positions
}
\description{
Generic function to extract movement track.
}
\seealso{
\code{\link{getTrack,Animal-method}} \code{\link{getTrack,Movement-method}} \code{\link{getTrack,Observation-method}}
}
