% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measurement.R
\name{Measurement}
\alias{Measurement}
\title{Construct a Measurement object}
\usage{
Measurement(model = c("n", "t", "sh"), logSdObs = c(0, 0), corObs = 0,
  logCorrection = matrix(c(0.6507, 0.8231, 2.3432, 2.0532, 3.8514, 2.9602,
  3.828, 3.0463, 4.4417, 3.6951, 5.5714, 5.5149, 8, 8), 2, 7),
  splineXlogSd = 0, knotPars = numeric(5), df = numeric(8), minDf = 3,
  nauticalObs = FALSE, ...)
}
\arguments{
\item{model}{Measurement model distribution}

\item{logSdObs}{Variance parameters on log-scale}

\item{corObs}{Correlation parameter}

\item{logCorrection}{Correction parameters between location classes}

\item{splineXlogSd}{Longitudinal variance parameter for spline model}

\item{knotPars}{Latitudinal variance parameters for spline model}

\item{df}{Degrees of freedom parameters for t distribution}

\item{minDf}{Minimal degrees of freedom}

\item{nauticalObs}{Transform observations from longitude/latitude to nautical miles?}

\item{...}{Not currently used}
}
\value{
a Measurement object
}
\description{
Construct a Measurement object
}
\examples{
d <- subadult_ringed_seal
obs <- Observation(lon = d$lon,
                   lat = d$lat,
                   dates = as.POSIXct(d$date,tz="GMT"),
                   locationclass = d$lc
                   )
meas <- Measurement(model="t")
mov <- RW(unique(obs$dates))
anim <- Animal(obs,mov,meas,"Subadult")
plot(anim)
}
\seealso{
\code{\link{Measurement-class}} \code{\link{Movement}} \code{\link{Observation}} \code{\link{Animal}}
}
\author{
Christoffer Moesgaard Albertsen
}
