% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observation.R
\docType{class}
\name{Observation-class}
\alias{Observation-class}
\alias{Observation}
\title{A Reference Class for argosTrack observations.}
\description{
A Reference Class for argosTrack observations.
}
\section{Fields}{

\describe{
\item{\code{lat}}{Latitude of observations}

\item{\code{lon}}{Longitude of observations}

\item{\code{dates}}{POSIXct vector of time stamps for observations}

\item{\code{dayOfYear}}{Day of year corresponding to dates}

\item{\code{qual}}{Factor of Argos location quality classes}

\item{\code{varModelCode}}{Integer vector of variance model codes (0, 1, 2, or 3)}

\item{\code{include}}{Logical vector of whether observations should be included in likelihood}
}}

\section{Methods}{

\describe{
\item{\code{addToLatPlot(arglist = list(pch = 16, col = "grey"))}}{Function to add observed latitudes to a plot. arglist is a list of base graphics arguments to be used for plotting.}

\item{\code{addToLonPlot(arglist = list(pch = 16, col = "grey"))}}{Function to add observed longitudes to a plot. arglist is a list of base graphics arguments to be used for plotting.}

\item{\code{addToMapPlot(arglist = list(type = "b", pch = 16, col = "grey"))}}{Function to add observed coordinates to a plot. arglist is a list of base graphics arguments to be used for plotting.}

\item{\code{getRange()}}{Returns the range of observed coordinates.}

\item{\code{getTMBdata()}}{Function to return a data list for TMB::MakeADFun}

\item{\code{getTMBmap(...)}}{Function to return a map list for TMB::MakeADFun.}

\item{\code{getTMBparameters()}}{Function to return a parameter list for TMB::MakeADFun}

\item{\code{initialize(lon, lat, dates, locationclass, include = rep(TRUE, length(dates)),
  ...)}}{Method to initialize the class. 'lon' is a vector of longitudes, 'lat' is a vector of latitudes, 'dates' is a vector of POSIXct dates, 'locationclass' is a factor of Argos locatin classes, and Observation reference class object, 'movement' is a Movement reference class object, 'measurement' is a Measuremenbt reference class object, and 'include' is a logical vector of whether observations should be included in likelihood.}
}}

\examples{
d <- subadult_ringed_seal
obs <- Observation(lon = d$lon,
                   lat = d$lat,
                   dates = as.POSIXct(d$date,tz="GMT"),
                   locationclass = d$lc
                   )
plot(obs)

}
\seealso{
\code{\link{Movement-class}}, \code{\link{Measurement}}, \code{\link{Animal}}
}
\author{
Christoffer Moesgaard Albertsen
}
