% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observation.R
\name{Observation}
\alias{Observation}
\title{Create an Observation object}
\usage{
Observation(
  lon,
  lat,
  dates,
  locationclass,
  include = rep(TRUE, length(dates)),
  ...
)
}
\arguments{
\item{lon}{Observed longitude (x-coordinate)}

\item{lat}{Observed latitude (y-coordinate)}

\item{dates}{Dates of observations}

\item{locationclass}{Location class of observations}

\item{include}{Vector indicating if each observation should be included.}

\item{...}{}
}
\value{
Observation reference class object
}
\description{
Create an Observation object
}
\author{
Christoffer Moesgaard Albertsen
}
