% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_OUF_.R
\docType{class}
\name{OUF-class}
\alias{OUF-class}
\title{A Reference Class for fitting an Ornstein-Uhlenbeck with foraging.}
\description{
The reference class implements an Ornstein-Uhlenbeck with foraging (Flemming et al. 2014).
}
\section{Methods}{

\describe{
\item{\code{getTMBmap(...)}}{Function to return a map list for TMB::MakeADFun.}

\item{\code{simulate(x0 = c(0, 0, 0, 0))}}{Function to simulate from the movement model. The initial states (latitudianl velocity, longitudinal velocity, latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}

\references{
Flemming et al.
}
\seealso{
\link{Movement-class} \link{OUF}

Other "Movement models": 
\code{\link{CSB-class}},
\code{\link{CTCRW-class}},
\code{\link{DCRW-class}},
\code{\link{DIRAC-class}},
\code{\link{DSBHN-class}},
\code{\link{DSBW-class}},
\code{\link{GDCRW-class}},
\code{\link{MPCTCRW-class}},
\code{\link{Movement-class}},
\code{\link{OUL-class}},
\code{\link{OUV-class}},
\code{\link{RW-class}}
}
\concept{"Movement models"}
