% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_OUF_.R
\name{OUF}
\alias{OUF}
\title{Create an OUF movement model object}
\usage{
OUF(
  dates,
  pars = numeric(6),
  varPars = numeric(2),
  nauticalStates = FALSE,
  timeunit = "hours"
)
}
\arguments{
\item{dates}{Vector of distinct and increasing POSIXct dates}

\item{pars}{Vector of movement parameters: \\eqn{log(\\beta_{lat})}, \\eqn{log(\\beta_{lon})}, \\eqn{\\gamma_{lat}}, \\eqn{\\gamma_{lon}}}

\item{varPars}{Vector of movement variance parameters: \\eqn{log(\\sigma_{lat})}, \\eqn{log(\\sigma_{lon})}}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
An OUF object
}
\description{
Create an OUF movement model object
}
\examples{
d <- subadult_ringed_seal
mov <- OUF(unique(as.POSIXct(d$date,tz="GMT")))
}
\seealso{
\code{\link{OUF-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
