% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_OUL.R
\name{OUL}
\alias{OUL}
\title{Create an OUL movement model object}
\usage{
OUL(
  dates,
  pars = c(1, 0, 0, 1, 0, 0),
  varPars = numeric(2),
  nauticalStates = FALSE,
  timeunit = "hours"
)
}
\arguments{
\item{dates}{Vector of distinct and increasing POSIXct dates}

\item{pars}{Vector of movement parameters: \\eqn{B_{1,1}}, \\eqn{B_{2,1}}, \\eqn{B_{1,2}}, \\eqn{B_{2,2}}, \\eqn{\\gamma_{lat}}, \\eqn{\\gamma_{lon}}}

\item{varPars}{Vector of movement variance parameters: \\eqn{log(\\sigma_{lat})}, \\eqn{log(\\sigma_{lon})}}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
A OUL object
}
\description{
Create an OUL movement model object
}
\examples{
d <- subadult_ringed_seal
mov <- OUL(unique(as.POSIXct(d$date,tz="GMT")))
}
\seealso{
\code{\link{OUL-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
