% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.argostrack}
\alias{plot.argostrack}
\title{Plot the fit from argosTrack}
\usage{
\method{plot}{argostrack}(x, bg_style = c("none", "map"), only_map = FALSE,
  min_area = 0.01, zoom_to_obs = TRUE, ...)
}
\arguments{
\item{x}{fitted argosTrack object}

\item{bg_style}{If 'none' an empty background is plotted. If 'map' a map is plotted in the background.}

\item{only_map}{Should only a map be plotted or also coordinate wise estimates?}

\item{min_area}{Minimum area of land polygons to plot.}

\item{zoom_to_obs}{If TRUE all observations will be shown. If FALSE outliers may be left outside the plotting area.}

\item{...}{}
}
\description{
Plot the fit from argosTrack
}
\author{
Christoffer Moesgaard Albertsen
}

