% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Movement}
\alias{plot.Movement}
\title{Plotting a Movement object}
\usage{
\method{plot}{Movement}(x, ...)
}
\arguments{
\item{x}{Movement reference class object}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
Invisibly returns the reference class object
}
\description{
Plots time \eqn{\times} latitude, time \eqn{\times} longitude and longitude \eqn{\times} latitude for a Movement reference class object.
}
\seealso{
\code{\link{plotMap}}, \code{\link{plotLat}}, and \code{\link{plotLon}}
}
\author{
Christoffer Moesgaard Albertsen
}
