% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Observation}
\alias{plot.Observation}
\title{Plotting an Observation object}
\usage{
\method{plot}{Observation}(x, ...)
}
\arguments{
\item{x}{Observation reference class object}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Plots time \eqn{\times} latitude, time \eqn{\times} longitude and longitude \eqn{\times} latitude for an Observation reference class object.
}
\author{
Christoffer Moesgaard Albertsen
}
\seealso{
\code{\link{plotMap}}, \code{\link{plotLat}}, and \code{\link{plotLon}}
}

