% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.summary_argostrack}
\alias{plot.summary_argostrack}
\title{Plot of the summary of an argosTrack fit.}
\usage{
\method{plot}{summary_argostrack}(x, nclass = 35, prob = TRUE,
  type = "both", bearings = FALSE, ...)
}
\arguments{
\item{x}{}

\item{nclass}{}

\item{prob}{}

\item{type}{}

\item{bearings}{}

\item{...}{}
}
\description{
Plot of the summary of an argosTrack fit.
}
\author{
Christoffer Moesgaard Albertsen
}

