% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotLat.R
\docType{methods}
\name{plotLat,Animal-method}
\alias{plotLat,Animal-method}
\title{Latitude plot for an Animal reference class object}
\usage{
\S4method{plotLat}{Animal}(object, plotArgs = list(), args = list(lwd = 3,
  col = "red"), add = FALSE, obsArgs = list(pch = 16), sdArgs = list(col =
  "grey", border = NA), sd = FALSE, ...)
}
\arguments{
\item{object}{Animal reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting latitude movement data.}

\item{add}{If FALSE a new plot window is created.}

\item{obsArgs}{Arguments for plotting latitude observation data.}

\item{sdArgs}{Arguments for plotting standard errors.}

\item{sd}{Should standard errors be plotted?}

\item{...}{additional arguments}
}
\description{
Latitude plot for an Animal reference class object
}
\author{
Christoffer Moesgaard Albertsen
}
\seealso{
\code{\link{plotLat}}, \code{\link{plotLat,Movement-method}}, \code{\link{plotLat,Observation-method}}
}

