% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLat.R
\docType{methods}
\name{plotLat,Movement-method}
\alias{plotLat,Movement-method}
\title{Latitude plot for a Movement reference class object}
\usage{
\S4method{plotLat}{Movement}(object, plotArgs = list(), args = list(),
  add = FALSE, sdArgs = list(col = "grey", border = NA), sd = FALSE, ...)
}
\arguments{
\item{object}{Movement reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting latitude movement data.}

\item{add}{If FALSE a new plot window is created.}

\item{sdArgs}{Arguments for plotting standard errors.}

\item{sd}{Should standard errors be plotted?}

\item{...}{additional arguments}
}
\description{
Latitude plot for a Movement reference class object
}
\seealso{
\code{\link{plotLat}}, \code{\link{plotLat,Animal-method}}, \code{\link{plotLat,Observation-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
