% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotLat.R
\docType{methods}
\name{plotLat,Observation-method}
\alias{plotLat,Observation-method}
\title{Latitude plot for an Observation reference class object}
\usage{
\S4method{plotLat}{Observation}(object, plotArgs = list(), args = list(),
  add = FALSE, ...)
}
\arguments{
\item{object}{Observation reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting latitude observation data.}

\item{add}{If FALSE a new plot window is created.}

\item{...}{additional arguments}
}
\description{
Latitude plot for an Observation reference class object
}
\author{
Christoffer Moesgaard Albertsen
}
\seealso{
\code{\link{plotLat}}, \code{\link{plotLat,Animal-method}}, \code{\link{plotLat,Movement-method}}
}

