% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotLat.R
\name{plotLat}
\alias{plotLat}
\title{Generic latitude plot}
\usage{
plotLat(object, ...)
}
\arguments{
\item{object}{Object to plot}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Generic latitude plot
}
\author{
Christoffer Moesgaard Albertsen
}
\seealso{
\code{\link{plotLat,Animal-method}}, \code{\link{plotLat,Movement-method}}, \code{\link{plotLat,Observation-method}}
}

