% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLon.R
\name{plotLon,Animal-method}
\alias{plotLon,Animal-method}
\title{Longitude plot for an Animal reference class object}
\usage{
\S4method{plotLon}{Animal}(
  object,
  plotArgs = list(),
  args = list(lwd = 3, col = "red"),
  add = FALSE,
  obsArgs = list(pch = 16),
  sdArgs = list(col = "grey", border = NA),
  sd = FALSE,
  ...
)
}
\arguments{
\item{object}{Animal reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting longitude movement data.}

\item{add}{If FALSE a new plot window is created.}

\item{obsArgs}{Arguments for plotting longitude observation data.}

\item{sdArgs}{Arguments for plotting standard errors.}

\item{sd}{Should standard errors be plotted?}

\item{...}{additional arguments}
}
\value{
Invisibly returns the reference class object
}
\description{
Longitude plot for an Animal reference class object
}
\seealso{
\code{\link{plotLon}}, \code{\link{plotLon,Movement-method}}, \code{\link{plotLon,Observation-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
