% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLon.R
\docType{methods}
\name{plotLon,Movement-method}
\alias{plotLon,Movement-method}
\title{Longitude plot for a Movement reference class object}
\usage{
\S4method{plotLon}{Movement}(object, plotArgs = list(), args = list(),
  add = FALSE, sdArgs = list(col = "grey", border = NA), sd = TRUE, ...)
}
\arguments{
\item{object}{Movement reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting longitude movement data.}

\item{add}{If FALSE a new plot window is created.}

\item{sdArgs}{Arguments for plotting standard errors.}

\item{sd}{Should standard errors be plotted?}

\item{...}{additional arguments}
}
\value{
Invisibly returns the reference class object
}
\description{
Longitude plot for a Movement reference class object
}
\seealso{
\code{\link{plotLon}}, \code{\link{plotLon,Animal-method}}, \code{\link{plotLon,Observation-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
