% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLon.R
\docType{methods}
\name{plotLon,Observation-method}
\alias{plotLon,Observation-method}
\title{Longitude plot for an Observation reference class object}
\usage{
\S4method{plotLon}{Observation}(object, plotArgs = list(),
  args = list(), add = FALSE, ...)
}
\arguments{
\item{object}{Observation reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting Longitude observation data.}

\item{add}{If FALSE a new plot window is created.}

\item{...}{additional arguments}
}
\value{
Invisibly returns the reference class object
 @seealso \code{\link{plotLon}}, \code{\link{plotLon,Animal-method}}, \code{\link{plotLon,Movement-method}}
}
\description{
Longitude plot for an Observation reference class object
}
\author{
Christoffer Moesgaard Albertsen
}
