% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotLon.R
\name{plotLon}
\alias{plotLon}
\title{Generic longitude plot}
\usage{
plotLon(object, ...)
}
\arguments{
\item{object}{Object to plot}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Generic longitude plot
}
\author{
Christoffer Moesgaard Albertsen
}
\seealso{
\code{\link{plotLon,Animal-method}}, \code{\link{plotLon,Movement-method}}, \code{\link{plotLon,Observation-method}}
}

