% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMap.R
\docType{methods}
\name{plotMap,Animal-method}
\alias{plotMap,Animal-method}
\title{Map plot for an Animal reference class object}
\usage{
\S4method{plotMap}{Animal}(object, plotArgs = list(), args = list(lwd = 3,
  col = "red"), add = FALSE, obsArgs = list(pch = 16), sdArgs = list(col =
  "grey", border = NA), ...)
}
\arguments{
\item{object}{Animal reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting movement data.}

\item{add}{If FALSE a new plot window is created.}

\item{obsArgs}{Arguments for plotting observation data.}

\item{sdArgs}{Arguments for plotting standard errors.}

\item{...}{additional arguments}
}
\description{
Map plot for an Animal reference class object
}
\author{
Christoffer Moesgaard Albertsen
}
\seealso{
\code{\link{plotMap}}, \code{\link{plotMap,Movement-method}}, \code{\link{plotMap,Observation-method}}
}

