% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\docType{methods}
\name{plotMap,Movement-method}
\alias{plotMap,Movement-method}
\title{Map plot for a Movement reference class object}
\usage{
\S4method{plotMap}{Movement}(object, plotArgs = list(), args = list(pch
  = 16, lwd = 3, col = "red"), add = FALSE, sdArgs = list(col = "grey",
  border = NA), ...)
}
\arguments{
\item{object}{Movement reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting movement data.}

\item{add}{If FALSE a new plot window is created.}

\item{sdArgs}{Arguments for plotting standard errors.}

\item{...}{additional arguments}
}
\value{
Invisibly returns the reference class object
}
\description{
Map plot for a Movement reference class object
}
\seealso{
\code{\link{plotMap}}, \code{\link{plotMap,Animal-method}}, \code{\link{plotMap,Observation-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
