% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\docType{methods}
\name{plotMap,Observation-method}
\alias{plotMap,Observation-method}
\title{Map plot for an Observation reference class object}
\usage{
\S4method{plotMap}{Observation}(object, plotArgs = list(),
  args = list(pch = 16, lwd = 3, col = "red"), add = FALSE,
  sdArgs = list(col = "grey", border = NA), ...)
}
\arguments{
\item{object}{Observation reference class object}

\item{plotArgs}{Arguments to setup background plot}

\item{args}{Arguments for plotting latitude observation data.}

\item{add}{If FALSE a new plot window is created.}

\item{sdArgs}{Arguments for plotting standard errors.}

\item{...}{additional arguments}
}
\value{
Invisibly returns the reference class object
}
\description{
Map plot for an Observation reference class object
}
\seealso{
\code{\link{plotMap}}, \code{\link{plotMap,Animal-method}}, \code{\link{plotMap,Movement-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
