% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.argostrack}
\alias{residuals.argostrack}
\title{Extract model residuals}
\usage{
\method{residuals}{argostrack}(object, type = "onestep", seed = 1,
  parallel = FALSE, osamethod = NULL, ...)
}
\description{
Extract model residuals
}

