% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample.R
\name{rhalfnorm}
\alias{rhalfnorm}
\title{Simulate from a half-normal distribution}
\usage{
rhalfnorm(n, sigma)
}
\arguments{
\item{n}{Number of replications}

\item{sigma}{Scale parameter}
}
\value{
Vector of simulated values
}
\description{
Simulates from a univariate half-normal distribution, \eqn{HN(\sigma^2)}. If \eqn{X\sim N(0,\sigma^2)} then \eqn{|X|\sim HN(\sigma^2)}.
}
\author{
Christoffer Moesgaard Albertsen
}
\keyword{internal}

