% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Simulate from a multivariate normal distribution}
\usage{
rmvnorm(n, mu, sigma)
}
\arguments{
\item{n}{Number of replications}

\item{mu}{Mean vector to simulate with}

\item{sigma}{Covariance matrix to simulate with}
}
\value{
A n x length(mu) matrix of simulated values
}
\description{
Simulates values from a \eqn{k}-variate normal distribution \eqn{N_k(\mu,\Sigma)} with density
\deqn{f(x) = \sqrt{(2\pi)^k\mid\Sigma\mid}^{-1}\exp\left( -\frac{1}{2}(x-\mu)^T\Sigma^{-1}(x-\mu)  \right)}
}
\author{
Christoffer Moesgaard Albertsen
}
\keyword{internal}

