% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample.R
\name{rmvt}
\alias{rmvt}
\title{Simulate from a multivariate Student's t-distribution}
\usage{
rmvt(n, mu, sigma, df)
}
\arguments{
\item{n}{Number of replications}

\item{mu}{Mean vector (\eqn{\mu}) to simulate with}

\item{sigma}{Scale matrix (\eqn{\Sigma}) to simulate with}

\item{df}{Degrees of freedom (\eqn{\nu}) to simulate with}
}
\value{
A n x length(mu) matrix of simulated values
}
\description{
Simulates values from a \eqn{k}-variate t-distribution \eqn{t_k(\mu,\Sigma,\nu)}. If \eqn{X \sim N_k(0,\Sigma)}, \eqn{v \sim \chi_\nu ^2}, and \eqn{Y = X\sqrt{\nu/v} + \mu} then
\deqn{Y \sim t_k(\mu,\Sigma,\nu)}
}
\author{
Christoffer Moesgaard Albertsen
}
\keyword{internal}

