% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roseplot_obj.R
\docType{methods}
\name{roseplot,Animal-method}
\alias{roseplot,Animal-method}
\title{Roseplot for Animal reference class object}
\usage{
\S4method{roseplot}{Animal}(object, ...)
}
\arguments{
\item{object}{Animal reference class object}

\item{...}{parameters passed to the method for the Movement reference class}
}
\value{
As \code{\link{roseplot,Movement-method}}
}
\description{
Calls roseplot on object$movement
}
\seealso{
\code{\link{roseplot,Movement-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
