% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roseplot_obj.R
\name{roseplot,Movement-method}
\alias{roseplot,Movement-method}
\title{Roseplot for Movement reference class object}
\usage{
\S4method{roseplot}{Movement}(
  object,
  nclass = 35,
  prob = TRUE,
  type = c("both", "step", "angle"),
  bearings = FALSE,
  show = TRUE,
  ...
)
}
\arguments{
\item{object}{Movement reference class object}

\item{nclass}{Number of histogram classes}

\item{prob}{Scale the frequencies to probabilities?}

\item{type}{'step' for plotting step lengths, 'angle' for plotting angles and 'both' for plotting both.}

\item{bearings}{Plot bearings (TRUE) or turning angles (FALSE)}

\item{show}{Should a plot be produced?}

\item{...}{not used}
}
\value{
Invisibly returns a list of bearings, step lengths and step lengths in km/h.
}
\description{
Roseplot for Movement reference class object
}
\examples{
d <- subadult_ringed_seal
obs <- Observation(lon = d$lon,
                   lat = d$lat,
                   dates = as.POSIXct(d$date,tz="GMT"),
                   locationclass = d$lc
                   )
meas <- Measurement(model="n")
mov <- RW(unique(obs$dates))
roseplot(mov)

}
\author{
Christoffer Moesgaard Albertsen
}
