% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roseplot_obj.R
\name{roseplot}
\alias{roseplot}
\title{Generic roseplot}
\usage{
roseplot(object, ...)
}
\arguments{
\item{object}{Object to plot}

\item{...}{other parameters}
}
\description{
Generic roseplot
}
\author{
Christoffer Moesgaard Albertsen
}
