% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Movement_RW.R
\docType{class}
\name{RW-class}
\alias{RW}
\alias{RW-class}
\title{A Reference Class for fitting a Random Walk model.}
\description{
The reference class implements a Random Walk (e.g. Nielsen et al. 2006).
\deqn{X_t \sim N(X_{t-\Delta_t},\Delta_t diag(\sigma_1^2,\sigma_2^2))}
}
\section{Methods}{

\describe{
\item{\code{initialize(dates, pars = numeric(0), varPars = numeric(2),
  nauticalStates = FALSE, timeunit = "hours")}}{Method to initialize the class. 'dates' is a vector of distinct and increasing POSIXct dates; 'pars' is zero length vector of the movement parameters; 'varPars' is a vector of movement variance parameters: \eqn{log(\sigma_{lat})}, \eqn{log(\sigma_{lon})}; 'nauticalStates' is a logical value indicating whether the states should be modelled in nautical miles, and 'timeunit' is the time unit to use for calculating time steps.}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. The function only returns the locations. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}
\examples{
d <- subadult_ringed_seal
mov <- CTCRW(unique(as.POSIXct(d$date)))
}
\references{
Nielsen, A., Bigelow, K. A., Musyl, M. K. and Sibert, J. R. (2006) Improving light-based geolocation by including sea surface temperature. Fisheries Oceanography, 15: 314-325. doi: 10.1111/j.1365-2419.2005.00401.x
}
\seealso{
\link{Movement}

Other X.Movement.models.: \code{\link{CTCRW}},
  \code{\link{CTCRW-class}}; \code{\link{DCRW}},
  \code{\link{DCRW-class}}; \code{\link{DSBHN}},
  \code{\link{DSBHN-class}}; \code{\link{DSBW}},
  \code{\link{DSBW-class}}; \code{\link{IDCRW}},
  \code{\link{IDCRW-class}}; \code{\link{MPCTCRW}},
  \code{\link{MPCTCRW-class}}; \code{\link{Movement}},
  \code{\link{Movement-class}}
}

