% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_RW.R
\name{RW}
\alias{RW}
\title{Create a RW movement model object}
\usage{
RW(dates, pars = numeric(0), varPars = numeric(2), nauticalStates = FALSE,
  timeunit = "hours")
}
\arguments{
\item{dates}{Vector of distinct and increasing POSIXct dates}

\item{pars}{Zero length vector of movement parameters}

\item{varPars}{Vector of movement variance parameters: \\eqn{log(\\sigma_{lat})}, \\eqn{log(\\sigma_{lon})}}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
A RW object
}
\description{
Create a RW movement model object
}
\examples{
d <- subadult_ringed_seal
mov <- RW(unique(as.POSIXct(d$date,tz="GMT")))
}
\seealso{
\code{\link{RW-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
