% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simTrack.R
\docType{methods}
\name{simTrack,Measurement-method}
\alias{simTrack,Measurement-method}
\title{Simulate from a measurement model with observation information}
\usage{
\S4method{simTrack}{Measurement}(object, n = 1, observation, ...)
}
\arguments{
\item{object}{Measurement reference class object to simulate from}

\item{n}{Number of replications}

\item{observation}{Observation reference class object with time points and Argos location class information to use in the simulation}

\item{...}{Not used}
}
\value{
A 2 x number of time steps in observation x n array of simulated values
}
\description{
Simulate from a measurement model with observation information
}
\author{
Christoffer Moesgaard Albertsen
}
\seealso{
\code{\link{simTrack}}, \code{\link{simTrack,Animal-method}}, \code{\link{simTrack,Movement-method}}
}

