% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrack.R
\name{simTrack,Movement-method}
\alias{simTrack,Movement-method}
\title{Simulate from a movement model}
\usage{
\S4method{simTrack}{Movement}(object, n = 1, x0 = object$mu[, 1], ...)
}
\arguments{
\item{object}{Movement reference class object implementing the model to simulate from}

\item{n}{Number of replications}

\item{x0}{Initial values}

\item{...}{not used}
}
\value{
A 2 x number of time steps in object x n array of simulated values
}
\description{
Simulate from a movement model
}
\seealso{
\code{\link{simTrack}}, \code{\link{simTrack,Animal-method}}, \code{\link{simTrack,Measurement-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
