% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrack.R
\name{simTrack}
\alias{simTrack}
\title{Generic function to simulate tracks}
\usage{
simTrack(object, n, ...)
}
\arguments{
\item{object}{Object to simulate from}

\item{n}{Number of replications}

\item{...}{other parameters}
}
\description{
Generic function to simulate tracks
}
\seealso{
\code{\link{simTrack,Animal-method}}, \code{\link{simTrack,Measurement-method}}, \code{\link{simTrack,Movement-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
