% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.argostrack}
\alias{simulate.argostrack}
\title{Simulate a new track from Argos data}
\usage{
\method{simulate}{argostrack}(object, nsim = 1, seed = NULL,
  locationclass = NULL, fixstates = FALSE, newpar = NULL, ...)
}
\arguments{
\item{object}{An argostrack object}

\item{nsim}{Number of simulations. Ignored.}

\item{seed}{Seed for simulations. Ignored.}

\item{locationclass}{Use other location classes than the object}

\item{fixstates}{Use other location classes than the object}

\item{newpar}{List of new parameter values to use}

\item{...}{Not used}
}
\value{
a list of simulated values. See details.
}
\description{
Simulate a new track from Argos data
}
\details{
The function returns a list of lists of simulated values containing:
\itemize{
\item{velocity}{Matrix of simulated coordinatewise velocities}
\item{position}{Matrix of true location coordinates}
\item{observation}{Matrix of observed location coordinates}
\item{dates}{Dates of observations}
\item{locationclass}{Location class of observations}
}
The simulations are based on a fitted argosTrack object.
}
\author{
Christoffer Moesgaard Albertsen
}

