% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{subadult_ringed_seal}
\alias{subadult_ringed_seal}
\title{Track data from the adult ringed seal with tag id 43847.}
\format{A data frame with 3586 rows and 5 variables:
\describe{
  \item{id}{Tag id}
  \item{date}{Date and time of the observation}
  \item{lc}{Argos location class of the observation}
  \item{lat}{Observed latitude (degrees north) from the Argos system}
  \item{lc}{Observed longitude (degrees east) from the Argos system}

}}
\source{
Fisheries and Oceans Canada
}
\usage{
subadult_ringed_seal
}
\description{
Track data from the adult ringed seal with tag id 43847.
}
\keyword{datasets}

