% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldShorelines}
\alias{worldShorelines}
\title{World shorelines in high resolution}
\format{A list of 145101 matrices each defining the longitude and latitude of a shoreline polygon. The polygons define the boundary between land and ocean, except Antarctica, and the boundary between Antarctica grounding-line and ocean.
The longitudes are between -180 and 180 degrees East, while the latitudes are between -90 and 90 degrees North.}
\source{
A Global Self-consistent, Hierarchical, High-resolution Geography Database \url{http://www.soest.hawaii.edu/pwessel/gshhg/index.html}
}
\usage{
worldShorelines
}
\description{
World shorelines in high resolution
}
\keyword{datasets}

