% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{worldShorelinesArea}
\alias{worldShorelinesArea}
\title{#' Area of world shoreline polygons in high resolution}
\format{A vector of length 145101 with areas of the World shoreline data polygons in the same order.}
\source{
A Global Self-consistent, Hierarchical, High-resolution Geography Database \url{http://www.soest.hawaii.edu/pwessel/gshhg/index.html}
}
\usage{
worldShorelinesArea
}
\description{
#' Area of world shoreline polygons in high resolution
}
\keyword{datasets}

