% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/argostrack.R
\name{argosTrack}
\alias{argosTrack}
\title{Estimate track from Argos data}
\usage{
argosTrack(lon, lat, dates, locationclass, include = rep(TRUE, length(dates)),
  equalbetas = TRUE, timevarybeta = 1, fixgammas = TRUE,
  fixcorrection = FALSE, dfVals = NULL, dfMap = NULL, minDf = 3,
  errordistribution = "t", movementmodel = "ctcrw", verbose = TRUE,
  timeunit = "mins", nStates = NULL, nauticalStates = FALSE,
  nauticalObs = FALSE, nlminb.control = list(eval.max = 2000, iter.max =
  1500, rel.tol = 0.001, x.tol = 0.015))
}
\arguments{
\item{lon}{}

\item{lat}{}

\item{dates}{}

\item{locationclass}{}

\item{include}{}

\item{equalbetas}{}

\item{fixgammas}{}

\item{fixcorrection}{}

\item{dfMap}{}

\item{errordistribution}{}

\item{verbose}{Write maximum gradient components to the terminal?}

\item{timeunit}{}

\item{modelcode}{}
}
\description{
Estimate animal movement track from Argos data using TMB (tmb-project.org).
}

