% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/argostrack.R
\name{argosTrack}
\alias{argosTrack}
\title{Estimate track from Argos data}
\usage{
argosTrack(lon, lat, dates, locationclass, include = rep(TRUE, length(dates)),
  equalbetas = TRUE, timevarybeta = 1, fixgammas = TRUE,
  fixcorrection = FALSE, dfVals = NULL, dfMap = NULL, minDf = 3,
  errordistribution = c("n", "t"), movementmodel = c("rw", "ctcrw",
  "mpctcrw", "dtcrw", "dsb"), silent = FALSE, timeunit = "hours",
  nStates = NULL, nauticalStates = FALSE, nauticalObs = FALSE,
  nlminb.control = list(eval.max = 2000, iter.max = 1500, rel.tol = 0.001,
  x.tol = 0.015))
}
\arguments{
\item{lon}{Vector of observed longitude coordinate.}

\item{lat}{Vector of observed latitude coordinate}

\item{dates}{Vector of dates of observations. Must be either numerical or characters convertable to POSIXct}

\item{locationclass}{Vector of Argos location class for observations. Values must be in c("3", "2", "1", "0", "A", "B","Z")}

\item{include}{Boolean vector indicating which observations to include in the analysis.}

\item{equalbetas}{Should the autocorrelation parameters be equal in the two coordinates?}

\item{timevarybeta}{Number of autocorrelation parameters over time.}

\item{fixgammas}{Should the drift parameters be equal in the two coordinates?}

\item{fixcorrection}{Should the observational variance be fixed relative to the best location class?}

\item{dfVals}{Specify the degrees of freedom used for each location class.}

\item{dfMap}{Empose hypotheses on the degrees of freedom.}

\item{minDf}{Lower bound for the degrees of freedom.}

\item{errordistribution}{Error distribution to use. Either multivariate Gaussian (n) or multivariate t (t).}

\item{movementmodel}{Movement model to use. See details.}

\item{silent}{Disable all tracing information from TMB?}

\item{timeunit}{Time unit to use when calculating time between states.}

\item{nStates}{Number of hidden states for \code{dtcrw} and \code{dsb}}

\item{nauticalStates}{Should the states be estimated in nautical miles?}

\item{nauticalObs}{Should the observations be transformed to nautical miles?}

\item{nlminb.control}{Control list passed to nlminb for estimation}
}
\value{
list of stuff
}
\description{
Estimate track from Argos data.
}
\details{
Function to estimate animal movement track from Argos data using TMB (\url{http://tmb-project.org}).
}
\examples{
dat <- subadult_ringed_seal
dat <- dat[!(dat$lc=="Z"),][1:100,]
args <- list(lon = dat$lon,
             lat = dat$lat,
             dates = as.character(dat$date),
             locationclass = dat$lc,
             silent = TRUE,
             fixcorrection=TRUE,
             errordistribution="n",
             movementmodel="rw"
         )
fit <- do.call("argosTrack",args)
}
\author{
Christoffer Moesgaard Albertsen
}

