% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_CSB.R
\docType{class}
\name{CSB-class}
\alias{CSB-class}
\title{A Reference Class for fitting a Continuous Time Step lengths / Bearings model.}
\description{
The reference class implements a Continuous Time Step lengths / Bearings model (Parton et al. 2016). The bearings are modelled by a Brownian Motion,
\deqn{B_{t} = B_{t-1}+\eta_{t},}
and the step lengths are modelled by an Ornstein-Uhlenbeck model,
\deqn{S_{t} = \mu + e^{-\beta_c\Delta_t}(S_{t-\Delta_t} - \mu) + \zeta_t.}
Here, \eqn{\Delta_t} is the length of the time steps and \eqn{\eta_{t}}, \eqn{\zeta_{t}} are random variables.
\eqn{\eta_{c,t}} follows a Gaussian distribution with mean zero and variance \eqn{V\left(\eta_{c,t}\right) = \sigma_c^2(1-e^{-2\beta_c\Delta_t})/(2\beta_c)}. The \eqn{\eta}'s are independent in time and coordiantes.
\eqn{\zeta_{t}} follows a Gaussian distribution with mean zero and variance
\deqn{V\left(\zeta_{t}\right)=\frac{\sigma_S^2}{2\beta}\left(1 - e^{2\beta\Delta_t}\right).}
\eqn{\eta_t} follows a Gaussian distribution with mean zero and variance
\deqn{V\left(\eta_t\right) = \Delta_t \sigma_B^2}
}
\section{Methods}{

\describe{
\item{\code{getTMBparameters()}}{Function to return a parameter list for TMB::MakeADFun}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. The function only returns the locations. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}

\references{
Parton, A., P. G. Blackwell, and A. Skarin. (2016) Bayesian Inference for Continuous Time Animal Movement Based on Steps and Turns. arXiv:1608.05583.
}
\seealso{
\link{Movement-class} \link{CSB}

Other "Movement models": 
\code{\link{CTCRW-class}},
\code{\link{DCRW-class}},
\code{\link{DIRAC-class}},
\code{\link{DSBHN-class}},
\code{\link{DSBW-class}},
\code{\link{GDCRW-class}},
\code{\link{MPCTCRW-class}},
\code{\link{Movement-class}},
\code{\link{OUF-class}},
\code{\link{OUL-class}},
\code{\link{OUV-class}},
\code{\link{RW-class}}
}
\concept{"Movement models"}
\keyword{internal}
