% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_DIRAC.R
\docType{class}
\name{DIRAC-class}
\alias{DIRAC-class}
\title{A Reference Class for fitting a Dirac movement model.}
\description{
The reference class implements a Dirac movement model with no movement
\deqn{X_t \sim \delta(X_0))}
}
\section{Methods}{

\describe{
\item{\code{getTMBmap(...)}}{Function to return a map list for TMB::MakeADFun.}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. The function only returns the locations. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}

\examples{
d <- subadult_ringed_seal
mov <- DIRAC(unique(as.POSIXct(d$date,tz="GMT")))

}
\seealso{
\link{Movement-class} \link{DIRAC}

Other "Movement models": 
\code{\link{CSB-class}},
\code{\link{CTCRW-class}},
\code{\link{DCRW-class}},
\code{\link{DSBHN-class}},
\code{\link{DSBW-class}},
\code{\link{GDCRW-class}},
\code{\link{MPCTCRW-class}},
\code{\link{Movement-class}},
\code{\link{OUF-class}},
\code{\link{OUL-class}},
\code{\link{OUV-class}},
\code{\link{RW-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
\concept{"Movement models"}
