% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_DIRAC.R
\name{DIRAC}
\alias{DIRAC}
\title{Create a DIRAC movement model object}
\usage{
DIRAC(dates, pars = numeric(0), varPars = numeric(0),
  nauticalStates = FALSE, timeunit = "hours")
}
\arguments{
\item{dates}{Vector of distinct and increasing POSIXct dates}

\item{pars}{Zero length vector of movement parameters}

\item{varPars}{Zero length vector of movement variance parameters}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
A DIRAC object
}
\description{
Create a DIRAC movement model object
}
\examples{
d <- subadult_ringed_seal
mov <- DIRAC(unique(as.POSIXct(d$date,tz="GMT")))
}
\seealso{
\code{\link{DIRAC-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
