% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_DSBW.R
\name{DSBW}
\alias{DSBW}
\title{Create a DSBW movement model object}
\usage{
DSBW(
  dates,
  pars = numeric(3),
  varPars = numeric(0),
  nauticalStates = FALSE,
  timeunit = "hours"
)
}
\arguments{
\item{dates}{Vector of distinct and increasing POSIXct dates}

\item{pars}{Vector of movement parameters: \\eqn{log(\\rho)}, \\eqn{log(shape)}, \\eqn{log(scale)}}

\item{varPars}{Zeero length vector of movement variance parameters}

\item{nauticalStates}{Should latent states be transformed from longitude/latitude to nautical miles?}

\item{timeunit}{timeunit used for calculating time steps.}
}
\value{
A DSBW object
}
\description{
Create a DSBW movement model object
}
\examples{
d <- subadult_ringed_seal
dates <- unique(as.POSIXct(d$date,tz="GMT"))
dseq <- seq(min(dates),max(dates), "day")
mov <- DSBW(dseq)
}
\seealso{
\code{\link{DSBW-class}}
}
\author{
Christoffer Moesgaard Albertsen
}
