% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expandMu}
\alias{expandMu}
\title{Expand estimated states to all time points}
\usage{
expandMu(eststat, dt)
}
\arguments{
\item{eststat}{2xn Matrix of estimated states}

\item{dt}{vector of length m>=n of time difference to previous observation}
}
\description{
Expand estimated states to all time points
}

