% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTrack.R
\name{getTrack}
\alias{getTrack}
\title{Generic function to extract movement track.}
\usage{
getTrack(object, ...)
}
\arguments{
\item{object}{The object from which to extract the track}

\item{...}{Additional arguments passed to the appropriate method}
}
\description{
Generic function to extract movement track.
}
\seealso{
\code{\link{getTrack,Animal-method}}
}

