% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_MPCTCRW.R
\docType{class}
\name{MPCTCRW-class}
\alias{MPCTCRW-class}
\title{A Reference Class for fitting a Multi-Persistence Continuous Time Correlated Random Walk model.}
\description{
This movement model extends \code{\link{CTCRW}} by having two velocity processes; one with short memory and one with longer memory.
}
\section{Methods}{

\describe{
\item{\code{getTMBmap(...)}}{Function to return a map list for TMB::MakeADFun.}

\item{\code{simulate(x0 = c(0, 0, 0, 0, 0, 0))}}{Function to simulate from the movement model. The initial states (latitudianl velocity 1, longitudinal velocity 1, latitudianl velocity 2, longitudinal velocity 2, latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. The function only returns the locations. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}

\examples{
d <- subadult_ringed_seal
mov <- argosTrack:::MPCTCRW(unique(as.POSIXct(d$date,tz="GMT")))


}
\seealso{
\link{Movement-class}

Other "Movement models": \code{\link{CSB-class}},
  \code{\link{CTCRW-class}}, \code{\link{DCRW-class}},
  \code{\link{DIRAC}}, \code{\link{DSBHN-class}},
  \code{\link{DSBW}}, \code{\link{GDCRW-class}},
  \code{\link{Movement-class}}, \code{\link{OUL-class}},
  \code{\link{OUV-class}}, \code{\link{RW-class}}
}
\keyword{internal}
