% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_OUL.R
\docType{class}
\name{OUL-class}
\alias{OUL-class}
\title{A Reference Class for fitting an Ornstein-Uhlenbeck Location model.}
\description{
The reference class implements an Ornstein-Uhlenbeck Location (Blackwell 2016). The locations follow an Ornstein-Uhlenbeck process
\deqn{X_t - \gamma = e^{B\Delta_t}[X_{t-\Delta_t} - \gamma] + \eta_{t}}
where \eqn{\eta_t} is a bivariate normal with zero mean and covariance \eqn{\pmatrix{\sigma_1^2 & 0 \cr 0 & \sigma_2^2}}.
}
\section{Methods}{

\describe{
\item{\code{getTMBmap(...)}}{Function to return a map list for TMB::MakeADFun.}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}

\references{
Blackwell, P. G., Niu, M., Lambert, M. S. and LaPoint, S. D. (2016), Exact Bayesian inference for animal movement in continuous time. Methods Ecol Evol, 7: 184-195. doi:10.1111/2041-210X.12460
}
\seealso{
\code{\link{Movement-class}}, \code{\link{OUL}}.

Other "Movement models": \code{\link{CSB-class}},
  \code{\link{CTCRW-class}}, \code{\link{DCRW-class}},
  \code{\link{DIRAC-class}}, \code{\link{DSBHN-class}},
  \code{\link{DSBW-class}}, \code{\link{GDCRW-class}},
  \code{\link{MPCTCRW-class}},
  \code{\link{Movement-class}}, \code{\link{OUV-class}},
  \code{\link{RW-class}}
}
\concept{"Movement models"}
