% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_OUV.R
\docType{class}
\name{OUV-class}
\alias{OUV-class}
\title{A Reference Class for fitting an Ornstein-Uhlenbeck Velocity model.}
\description{
The reference class implements an Ornstein-Uhlenbeck Velocity. The velocity is modelled by the (bivariate) stochastic differential equation
\deqn{dV_t = -\Theta (V_t-\mu) d_t + S dB_t,}
Since the locations are know from the velocity by \eqn{X_{t} = \sum_{i=0}^{t} V_i}, the increments are known as
\deqn{X_{t_n} - X_{t_{n-1}} = V_{t_n}.}
Hence,
\deqn{X_{t_n} = X_{t_{n-1}} + \mu + \Theta (X_{t_{n-1}} - X_{t_{n-2}} - \mu)}
}
\section{Methods}{

\describe{
\item{\code{getTMBmap(...)}}{Function to return a map list for TMB::MakeADFun.}

\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}

\examples{
d <- subadult_ringed_seal
#mov <- argosTrack:::OUV(unique(as.POSIXct(d$date,tz="GMT")))

}
\references{
Albertsen, C.M. (2016) Personal communication.
}
\seealso{
\code{\link{Movement-class}}. A generalization of \code{\link{DCRW}}, \code{\link{RW}}, and \code{\link{CTCRW}} (except only the location is modelled here and it is assumed to be known given the velocities)

Other "Movement models": 
\code{\link{CSB-class}},
\code{\link{CTCRW-class}},
\code{\link{DCRW-class}},
\code{\link{DIRAC-class}},
\code{\link{DSBHN-class}},
\code{\link{DSBW-class}},
\code{\link{GDCRW-class}},
\code{\link{MPCTCRW-class}},
\code{\link{Movement-class}},
\code{\link{OUF-class}},
\code{\link{OUL-class}},
\code{\link{RW-class}}
}
\concept{"Movement models"}
\keyword{internal}
