% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.summary_argostrack}
\alias{plot.summary_argostrack}
\title{Plot of the summary of an argosTrack fit.}
\usage{
\method{plot}{summary_argostrack}(x, nclass = 35, prob = TRUE,
  type = "both", bearings = FALSE, ...)
}
\arguments{
\item{x}{argosTrack object}

\item{nclass}{Number of classes in the histogram}

\item{prob}{Standardize to proportions?}

\item{type}{Type of plot: step, angle or both}

\item{bearings}{Plot bearings instead of turning angles?}

\item{...}{Passed to histogram and roseplot}
}
\description{
Plot of the summary of an argosTrack fit.
}
\author{
Christoffer Moesgaard Albertsen
}

