% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLat.R
\name{plotLat}
\alias{plotLat}
\title{Generic latitude plot}
\usage{
plotLat(object, ...)
}
\arguments{
\item{object}{Object to plot}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
Invisibly returns the reference class object
}
\description{
Generic latitude plot
}
\seealso{
\code{\link{plotLat,Animal-method}}, \code{\link{plotLat,Movement-method}}, \code{\link{plotLat,Observation-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
