% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plotMap}
\alias{plotMap}
\title{Generic map plot}
\usage{
plotMap(object, ...)
}
\arguments{
\item{object}{Object to plot}

\item{...}{other parameters to be passed through to plotting functions.}
}
\value{
Invisibly returns the reference class object
}
\description{
Generic map plot
}
\seealso{
\code{\link{plotMap,Animal-method}}, \code{\link{plotMap,Movement-method}}, \code{\link{plotMap,Observation-method}}
}
\author{
Christoffer Moesgaard Albertsen
}
