% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{roseplot}
\alias{roseplot}
\title{Roseplot / radar histogram}
\usage{
roseplot(x, breaks = "Sturges", prob = TRUE, main = NULL, xlab = "",
  ...)
}
\arguments{
\item{x}{Vector of values}

\item{breaks}{Either \code{Sturges} or the number of breaks to use.}

\item{prob}{Should the bars be standardized?}

\item{main}{Main title of the plot}

\item{xlab}{x-axis label of the plot}

\item{...}{Passed to polygons for plotting the bars}
}
\description{
Roseplot / radar histogram
}
\author{
Christoffer Moesgaard Albertsen
}
\keyword{internal}

