% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movement_RW.R
\docType{class}
\name{RW-class}
\alias{RW-class}
\title{A Reference Class for fitting a Random Walk model.}
\description{
The reference class implements a Random Walk (e.g. Nielsen et al. 2006).
\deqn{X_t \sim N(X_{t-\Delta_t},\Delta_t diag(\sigma_1^2,\sigma_2^2))}
}
\section{Methods}{

\describe{
\item{\code{simulate(x0 = c(0, 0))}}{Function to simulate from the movement model. The initial states (latitudinal/y-coordinate location and longitudinal/x-coordinate location) must be given. The function only returns the locations. If nauticalStates==TRUE, the result is returned in nautical miles.}
}}

\examples{
d <- subadult_ringed_seal
mov <- RW(unique(as.POSIXct(d$date,tz="GMT")))

}
\references{
Nielsen, A., Bigelow, K. A., Musyl, M. K. and Sibert, J. R. (2006) Improving light-based geolocation by including sea surface temperature. Fisheries Oceanography, 15: 314-325. doi: 10.1111/j.1365-2419.2005.00401.x
}
\seealso{
\link{Movement-class} \link{RW}

Other "Movement models": \code{\link{CSB-class}},
  \code{\link{CTCRW-class}}, \code{\link{DCRW-class}},
  \code{\link{DIRAC-class}}, \code{\link{DSBHN-class}},
  \code{\link{DSBW-class}}, \code{\link{GDCRW-class}},
  \code{\link{MPCTCRW-class}},
  \code{\link{Movement-class}}, \code{\link{OUL-class}},
  \code{\link{OUV-class}}
}
\concept{"Movement models"}
