% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\name{simStates}
\alias{simStates}
\title{Simulate states given an argostrack object}
\usage{
simStates(object, newpar = NULL)
}
\arguments{
\item{object}{Argostrack object}

\item{locationclass}{The location classes of the track}
}
\description{
Simulate states given an argostrack object
}

