% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate.argostrack}
\alias{simulate.argostrack}
\title{Simulate a new track from Argos data}
\usage{
\method{simulate}{argostrack}(object)
}
\arguments{
\item{object}{An argostrack object}
}
\description{
Simulate a new track from Argos data
}

