% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.argostrack}
\alias{simulate.argostrack}
\title{Simulate a new track from Argos data}
\usage{
\method{simulate}{argostrack}(object, locationclass = NULL,
  fixstates = FALSE, newpar = NULL, ...)
}
\arguments{
\item{object}{An argostrack object}

\item{locationclass}{Use other location classes than the object}

\item{fixstates}{Use the states from object}
}
\description{
Simulate a new track from Argos data
}

